Enable website app and move Map related views
